+++ evo1/atb/__init__.py
@@ -1,11 +1,18 @@
-from evo1.atb.base import EncounterID
-from evo1.atb.farming import FarmingGoal, SeqATBmove2D, calc_next_encounter
+from evo1.atb.base import SeqATBCombat
+from evo1.atb.encounter import Encounter, EncounterID, calc_next_encounter
+from evo1.atb.entity import atb_stats_from_memory
+from evo1.atb.farming import FarmingGoal, SeqATBmove2D
 from evo1.atb.manual import SeqATBCombatManual
+from evo1.atb.predict import predict_attack
 
 __all__ = [
+    "Encounter",
     "EncounterID",
     "SeqATBmove2D",
     "FarmingGoal",
     "calc_next_encounter",
+    "predict_attack",
+    "SeqATBCombat",
     "SeqATBCombatManual",
+    "atb_stats_from_memory",
 ]