+++ evo1/checkpoints.py
@@ -1,9 +1,28 @@
-from evo1.seq import MeadowFight, NoriaMines, OverworldToMeadow
+from engine.seq import SeqList
+from evo1.seq import (
+    CrystalCavern,
+    MeadowFight,
+    NoriaMines,
+    OverworldToCavern,
+    OverworldToMeadow,
+)
+
+
+class CavernCheckpoint(SeqList):
+    def __init__(self):
+        super().__init__(
+            name="Caverns experimental",
+            children=[
+                OverworldToCavern(),
+                CrystalCavern(),
+            ],
+        )
 
 
 def Checkpoints():
     return {
         "overworld_expr": OverworldToMeadow(),
         "meadow_expr": MeadowFight(),
+        "caverns_expr": CavernCheckpoint(),
         "noria_expr": NoriaMines(),
     }