+++ evo1/seq/__init__.py
@@ -8,10 +8,8 @@ from evo1.seq.overworld import (
     OverworldToNoria,
 )
 from evo1.seq.papurika import MeadowFight, PapurikaVillage
-from evo1.seq.start import Evoland1StartGame
 
 __all__ = [
-    "Evoland1StartGame",
     "Edel1",
     "OverworldToMeadow",
     "MeadowFight",